create release bash script for automation, remove some old temp files, refactor util to outside business package
