--- android/sdk/notes.md
@@ -1,1 +1,0 @@
-- Add gson as new dependency
