+++ android/sdk/src/main/java/io/o2mc/sdk/business/BatchGenerator.java
@@ -4,6 +4,7 @@ import android.util.Log;
 
 import java.util.List;
 
+import io.o2mc.sdk.Util;
 import io.o2mc.sdk.domain.Batch;
 import io.o2mc.sdk.domain.DeviceInformation;
 import io.o2mc.sdk.domain.Event;
